+++ CleanCode.SolidPrinciples/SingleResponsibility/Laugher.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Text.RegularExpressions;
+using System.Threading.Tasks;
+
+namespace CleanCode.SolidPrinciples
+{
+    class Laugher
+    {
+        private string _laughSound;
+
+        public Laugher(string laugh)
+        {
+            _laughSound = laugh;
+        }
+
+        public string GenerateLaugh(int nrOfTimes)
+        {
+            return string.Concat(Enumerable.Repeat(_laughSound, nrOfTimes));
+        }
+    }
+}
