+++ CleanCode.SolidPrinciples/SingleResponsibility/ReferencesCounter.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Text.RegularExpressions;
+using System.Threading.Tasks;
+
+namespace CleanCode.SolidPrinciples
+{
+    class ReferencesCounter
+    {
+        private string[] _references;
+
+        public ReferencesCounter(params string[] references)
+        {
+            _references = references;
+        }
+
+        public int CountReferences(string inputString)
+        {
+            return _references.Sum(r => Regex.Matches(inputString.ToLower(), Regex.Escape(r.ToLower())).Count);
+        }
+    }
+}
