+++ CleanCode.SolidPrinciples/SingleResponsibility/StringHolidaysUtils.cs
@@ -7,28 +7,13 @@ using System.Threading.Tasks;
 
 namespace CleanCode.SolidPrinciples
 {
-    class StringHolidaysUtils
+    class Holiday
     {
-        private string _laughSound;
-        private string[] _references;
-
         public string HolidayName { get; private set; }
 
-        public StringHolidaysUtils(string holidayName, string laugh, params string[] references)
+        public Holiday(string holidayName)
         {
-            _laughSound = laugh;
             HolidayName = holidayName;
-            _references = references;
-        }
-
-        public string GenerateLaugh(int nrOfTimes)
-        {
-            return string.Concat(Enumerable.Repeat(_laughSound, nrOfTimes));
-        }
-
-        public int CountReferences(string inputString)
-        {
-            return _references.Sum(r => Regex.Matches(inputString.ToLower(), Regex.Escape(r.ToLower())).Count);
         }
     }
 }
